// packages/core/src/config/types/VitalsUpdateFrequency.ts
/*
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2016-Present Datadog, Inc.
 */
export enum VitalsUpdateFrequency {
    FREQUENT = 'FREQUENT',
    AVERAGE = 'AVERAGE',
    RARE = 'RARE',
    NEVER = 'NEVER'
}